+++ packages/components/src/local/collab-status-board/index.spec.tsx
@@ -5,7 +5,7 @@ import renderer from 'react-test-renderer'
 import { ForceRole, MessageCustom, Role, ChannelCollab } from '@serge/custom-types'
 import { CUSTOM_MESSAGE, CollaborativeMessageStates, CHANNEL_COLLAB, InitialStates, CollaborativePermission, PARTICIPANT_COLLAB } from '@serge/config'
 import { forces, MessageTemplatesMockByKey, WargameMock } from '@serge/mocks'
-import CoaStatusBoard from './index'
+import CollabStatusBoard from './index'
 import moment from 'moment'
 
 /* Import mock data */
@@ -213,7 +213,7 @@ it('CoaStatusBoard renders correctly', () => {
 
   const tree = renderer
     .create(
-      <CoaStatusBoard
+      <CollabStatusBoard
         currentWargame='wargame-test'
         templates={MessageTemplatesMockByKey}
         forces={forces}
