+++ packages/components/src/local/collab-status-board/index.tsx
@@ -23,7 +23,7 @@ export interface ForceColor {
 }
 
 /* Render component */
-export const CoaStatusBoard: React.FC<Props> = ({
+export const CollabStatusBoard: React.FC<Props> = ({
   templates, messages, channelColb, isObserver, isUmpire, onChange, role, forces,
   gameDate, onMessageRead, onMarkAllAsRead, currentWargame
 }: Props) => {
@@ -149,4 +149,4 @@ export const CoaStatusBoard: React.FC<Props> = ({
   )
 }
 
-export default CoaStatusBoard
+export default CollabStatusBoard
