+++ packages/components/src/local/collab-status-board/types/props.d.ts
