+++ packages/components/src/local/molecules/collab-message-detail/README.md
