+++ packages/components/src/local/molecules/collab-message-detail/index.tsx
@@ -116,7 +116,7 @@ const injectFeedback = (message: MessageCustom, verb: string, feedback: string, 
 }
 
 /* Render component */
-export const ChannelCoaMessageDetail: React.FC<Props> = ({
+export const CollabMessageDetail: React.FC<Props> = ({
   templates, message, state, onChange, isObserver, isUmpire,
   role, channelColb, permission, assignees = [], collapseMe, gameDate, onRead, isReaded
 }) => {
@@ -423,4 +423,4 @@ export const ChannelCoaMessageDetail: React.FC<Props> = ({
   }
 }
 
-export default ChannelCoaMessageDetail
+export default CollabMessageDetail
