+++ packages/components/src/local/molecules/collab-message-detail/styles.module.scss
