+++ packages/components/src/local/molecules/collab-message-detail/types/props.d.ts
